Use COA; add 'output' option
