+++ bin/compare-minimizers
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../lib/index.js')
+require('../lib/cli.js')
